requirements.txt
requests~=2.31.0
tqdm~=4.66.1
pillow~=10.2.0
pyvis~=0.3.2
networkx~=3.2.1
